[MUNOVA-49] :recycle: refactor: 쿠폰 적용 로직 추가

+++ src/main/java/com/space/munova/order/dto/GetOrderDetailResponse.java
@@ -15,7 +15,7 @@ public record GetOrderDetailResponse (
         String userRequest,
         OrderStatus status,
         Long originPrice,
-        int discountPrice,
+        Long discountPrice,
         Long totalPrice,
         LocalDateTime orderDate,
         String paymentReceipt,
